social-app/requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.4.1
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==38.0.1
dataclasses==0.8
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.83.0
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.1
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.42
starlette==0.19.1
typing_extensions==4.1.1
ujson==4.0.1
urllib3==1.26.12
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
